Can't move into target while tutorial
Now i have to make it so that you can’t die during the tutorial . I
also have to make it so that you can drag to the target at the end of
the tutorial. Also, in the tutorial, show that you start with 15 coins
with an arrow pointing to the coins.